move HibikiModule to its own file.  Add setTimeout, setInterval, and clearInterval handlers
