+++ src/modules.ts
@@ -1,6 +1,6 @@
 // Copyright 2021-2022 Dashborg Inc
 
-import {isObject, unpackPositionalArgs, stripAtKeys, getHibiki, fullPath, getSS, setSS, smartEncodeParam, unpackArg, unpackAtArgs, base64ToArray, callHook} from "./utils";
+import {isObject, unpackPositionalArgs, stripAtKeys, getHibiki, fullPath, getSS, setSS, smartEncodeParam, unpackArg, unpackAtArgs, base64ToArray, callHook, validateModulePath} from "./utils";
 import {sprintf} from "sprintf-js";
 import type {HibikiState} from "./state";
 import type {FetchHookFn, Hibiki, HibikiAction, HandlerPathType, HibikiExtState, HttpConfig, HibikiActionString, HibikiVal, HibikiValObj} from "./types";
@@ -282,20 +282,6 @@ function makeFetchInit(req : HibikiRequest, url : URL, opts : HttpConfig) : Reco
     return fetchInit;
 }
 
-function validateModulePath(modName : string, hpath : HandlerPathType) {
-    if (hpath.url.startsWith("http://") || hpath.url.startsWith("https://") || hpath.url.startsWith("//") || !urlSameOrigin(hpath.url)) {
-        throw new Error(sprintf("Invalid %s module URL, must not specify a protocol or host: %s", modName, fullPath(hpath)));
-    }
-    if (!hpath.url.startsWith("/")) {
-        throw new Error(sprintf("Invalid %s module URL, must not specify a relative url: %s", modName, fullPath(hpath)));
-    }
-}
-
-function urlSameOrigin(urlStr : string) {
-    let url = new URL(urlStr, window.location.href);
-    return url.origin == window.location.origin;
-}
-
 function isActionStr(v : any) : boolean {
     if (v == null) {
         return false;
@@ -464,118 +450,4 @@ class LibModule {
     }
 }
 
-class HibikiModule {
-    state : HibikiState;
-
-    constructor(state : HibikiState, config : any) {
-        this.state = state;
-    }
-
-    callHandler(req : HibikiRequest) : Promise<any> {
-        validateModulePath("hibiki", req.callpath);
-        let handlerName = req.callpath.url;
-        if (handlerName == "/get-session-storage") {
-            return this.getSS(req);
-        }
-        else if (handlerName == "/set-session-storage") {
-            return this.setSS(req);
-        }
-        else if (handlerName == "/set-title") {
-            return this.setTitle(req);
-        }
-        else if (handlerName == "/update-url") {
-            return this.updateUrl(req);
-        }
-        else if (handlerName == "/sleep") {
-            return this.sleep(req);
-        }
-        else if (handlerName == "/navigate") {
-            return this.navigate(req);
-        }
-        else {
-            throw new Error("Invalid Hibiki Module handler: " + fullPath(req.callpath));
-        }
-    }
-
-    sleep(req : HibikiRequest) : Promise<any> {
-        let sleepMs = unpackArg(req.data, "ms", 0);
-        if (sleepMs == null || typeof(sleepMs) !== "number") {
-            throw new Error("sleep requires 'ms' parameter (must be number)");
-        }
-        return new Promise((resolve, reject) => {
-            setTimeout(() => resolve(true), sleepMs);
-        });
-    }
-
-    getSS(req : HibikiRequest) : Promise<any> {
-        let key = unpackArg(req.data, "key", 0);
-        if (key == null) {
-            throw new Error("get-session-storage requires 'key' parameter");
-        }
-        return getSS(key);
-    }
-
-    setSS(req : HibikiRequest) : Promise<any> {
-        let key = unpackArg(req.data, "key", 0);
-        if (key == null) {
-            throw new Error("set-session-storage requires 'key' parameter");
-        }
-        let val = DataCtx.demobx(unpackArg(req.data, "value", 1));
-        setSS(key, val);
-        return null;
-    }
-
-    setTitle(req : HibikiRequest) : Promise<any> {
-        let title = unpackArg(req.data, "title", 0);
-        if (title == null) {
-            throw new Error("set-title requires 'title' parameter");
-        }
-        document.title = String(title);
-        req.setData("$state.title", document.title);
-        return null;
-    }
-
-    updateUrl(req : HibikiRequest) : Promise<any> {
-        let data = stripAtKeys(req.data);
-        let {path: urlStr, raw: isRaw, replace, title} = unpackAtArgs(req.data);
-        let url = new URL((urlStr ?? ""), window.location.href);
-        for (let key in data) {
-            let val = data[key];
-            if (val == null) {
-                url.searchParams.delete(key);
-                continue;
-            }
-            url.searchParams.set(key, smartEncodeParam(val, isRaw));
-        }
-        if (title != null) {
-            document.title = title;
-        }
-        if (replace) {
-            window.history.replaceState(null, document.title, url.toString());
-        }
-        else {
-            window.history.pushState(null, document.title, url.toString());
-        }
-        hibikiState(req.state).setStateVars();
-        return null;
-    }
-
-    navigate(req : HibikiRequest) : Promise<any> {
-        let data = stripAtKeys(req.data);
-        let {path: urlStr, raw: isRaw} = unpackAtArgs(req.data);
-        let url = new URL((urlStr ?? ""), window.location.href);
-        for (let key in data) {
-            let val = data[key];
-            if (val == null) {
-                url.searchParams.delete(key);
-                continue;
-            }
-            url.searchParams.set(key, smartEncodeParam(val, isRaw));
-        }
-        // @ts-ignore (you definitely can set window.location to a string)
-        window.location = url.toString();
-        return null;
-    }
-}
-
-export {LocalModule, HttpModule, LibModule, HibikiModule};
+export {LocalModule, HttpModule, LibModule};
